support optionally validating SGXType
This patch lays the groundwork for validating extensions inside the PCK
certificate beyond what's needed to verify that a quote is valid. This
patch adds an optional validation option to check the SGX type. This
can be used to differentiate platforms with logical and cryptographic
integrity. In the future, we could validate other fields e.g. "SMT
Enabled".