Enhance IsSupported to further check provider
There are some attributes in configfs-tsm reports that are available on
some providers and not available on others. The IsSupported function
should check that the guest is on TDX, and not just that they have tsm
report support.

The newline in the comparison is due to the provider attribute's show
method adding a \n after the name.